requirements.txt
<<<<<<< HEAD
openpyxl==3.1.5
pandas==2.2.3
Pillow==10.4.0
plotly==5.24.1
streamlit==1.39.0
=======
openpyxl==3.1.5
pandas==2.2.3
Pillow==10.4.0
plotly==5.24.1
streamlit==1.39.0
>>>>>>> aee5db8 (Reorganize project into folders (data, assets, pages))
